+++ program/Cargo.toml
@@ -10,9 +10,8 @@ sp1-zkvm = { git = "https://github.com/succinctlabs/sp1.git", branch = "dev" }
 ed25519-consensus = { version = "2", default-features = false }
 blake2 = "0.10.6"
 sha2 = "0.10.8"
-alloy-primitives = { version = "0.7.5", features=["serde"] }
+alloy-primitives = { version = "0.7.5", features = ["serde"] }
 alloy-sol-types = "0.7.5"
-hex = "0.4.3"
 
 [patch.crates-io]
 ed25519-consensus = { git = "https://github.com/sp1-patches/ed25519-consensus", branch = "patch-v2.1.0" }
